Localstorage followup (#7)
* New task: repeat and grouping

* Allow for float arguments

* Update app info

* Add some extra local storage timestamps

* Done: localstorage followup

+++ [library]/Turtle.js
@@ -267,7 +267,7 @@ class Turtle {
 		const argArray = argString.split(',');
 
 		const result = argArray.map(
-			(element) => { return Number.parseInt(element); }
+			(element) => { return Number.parseFloat(element); }
 		);
 		return result;
 	}
